+++ NEWS.md
@@ -7,6 +7,27 @@ New features(Analysis):
 + Infer that `new $x` is of the template type `T` if `$x` is `class-string<T>` (#2257)
 + Improve detection of invalid arguments in code implicitly calling `__invoke`.
 + Support extracting template types from more forms of `callable` types. (#2264)
++ Support `@phan-assert`, `@phan-assert-true-condition`, and `@phan-assert-false-condition`.
+  Examples of side effects when this annotation is used on a function/method declaration:
+
+  - `@phan-assert int $x` will assert that the argument to the parameter `$x` is of type `int`.
+  - `@phan-assert !false $x` will assert that the argument to the parameter `$x` is not false.
+  - `@phan-assert !\Traversable $x` will assert that the argument to the parameter `$x` is not `Traversable` (or a subclass)
+  - `@phan-assert-true-condition $x` will make Phan infer that the argument to parameter `$x` is truthy if the function returned successfully.
+  - `@phan-assert-false-condition $x` will make Phan infer that the argument to parameter `$x` is falsey if the function returned successfully.
+  - This can be used in combination with Phan's template support.
+
+  See [tests/plugin_test/src/072_custom_assertions.php](tests/plugin_test/src/072_custom_assertions.php) for example uses of these annotations.
+
+Plugins:
+- Add `PHPUnitAssertionPlugin`.
+  This plugin will make Phan infer side effects from some of the uses of the helper methods PHPUnit provides within test cases.
+
+  - Infer that a condition is truthy from `assertTrue()` and `assertNotFalse()` (e.g. `assertTrue($x instanceof MyClass)`)
+  - Infer that a condition is null/not null from `assertNull()` and `assertNotNull()`
+  - Infer class type of `$actual` from `assertInstanceOf(MyClass::class, $actual)`
+  - Infer that `$actual` has the exact type of `$expected` after calling `assertSame($expected, $actual)`
+  - Other methods aren't supported yet.
 
 30 Dec 2018, Phan 1.1.10
 ------------------------